RenderBotFresh/requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
pandas_ta==0.3.14b0
requests==2.32.3
pydantic==2.8.2
